Replace streamLogical to cut RAM usage on larger tables
We shouldn't reify the whole stripe as rows early, as the extra
copies of Boxed structures really kill us on RAM usage.

This is a little bit slower on structures with a small number of
rows, but a little bit faster on those will larges number per
stripe. And on the demo-12-zlib.orc example, which is almost
2M rows in a single stripe, it halves the amount or RAM we
require to gather it.